+++ src/api/http/client.ts
@@ -1,16 +1,10 @@
-import {
-  ApolloClient,
-  InMemoryCache,
-  NormalizedCacheObject,
-} from '@apollo/client/core'
+import { HttpClient } from '@/types'
+import { ApolloClient, InMemoryCache } from '@apollo/client/core'
 
 /**
  * Creates an authenticated GQL HTTP client.
  */
-const createHttpClient = (
-  endpoint: string,
-  apiToken: string,
-): ApolloClient<NormalizedCacheObject> => {
+const createHttpClient = (endpoint: string, apiToken: string): HttpClient => {
   return new ApolloClient({
     uri: endpoint,
     cache: new InMemoryCache(),
