+++ src/utils/bytes2utf8.ts
@@ -1,0 +1,5 @@
+const bytes2utf8 = (b: ArrayBuffer): string => {
+  return Buffer.from(b).toString('utf-8')
+}
+
+export default bytes2utf8
