+++ src/utils/get-env.ts
@@ -1,0 +1,6 @@
+const getEnv = (key: string, defaultValue: string) => {
+  const env = process.env[key]
+  return env ?? defaultValue
+}
+
+export default getEnv
