+++ src/utils/parse-project-ids.ts
@@ -1,0 +1,6 @@
+const parseProjectIds = (id: string): string[] => {
+  // @TODO: This needs better error handling.
+  return id.split(',')
+}
+
+export default parseProjectIds
