+++ src/utils/require-env.ts
@@ -1,0 +1,9 @@
+const requireEnv = (key: string) => {
+  const env = process.env[key]
+  if (env === undefined) {
+    throw new Error(`Environment variable "${key}" is required`)
+  }
+  return env
+}
+
+export default requireEnv
