+++ src/vector/sources.ts
@@ -1,0 +1,6 @@
+const STDIN = `
+[sources.in]
+type = "stdin"
+`
+
+export { STDIN }
